+++ src/Electron/src/Main/Auth/GitLabApi.fs
@@ -56,6 +56,7 @@ let private toAuthUserDto (baseUrl: string) (user: ComponentsGitLabApi.CurrentUs
         Username = user.username
         Name = user.name
         Email = user.email
+        CommitEmail = user.commit_email
         AvatarUrl = user.avatar_url |> Option.defaultValue ""
         TargetDataHub = baseUrl
     }